plugins/ndf/skills/playwright-scenario-test/pyproject.toml
[project]
name = "playwright-scenario-test"
version = "0.3.0"
description = "Playwright + curl による Web シナリオ E2E テストの汎用フレームワーク (Claude Code skill)。Locator-first / web-first assertion / page_role 別 a11y + CWV 自動付加 / HUD 字幕付き動画。"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.50,<2.0",
    "pyyaml>=6.0",
    # webm → mp4 変換用。Playwright 録画は webm のみ生成するが、
    # Google Drive のプレビュアでは mp4 (H.264) の方が安定するため変換する。
    "imageio-ffmpeg>=0.4",
]

[project.optional-dependencies]
# Drive 連携: scripts/trace_link.py / build_gdoc_with_drive_links.py 等が使用
drive = [
    "google-api-python-client>=2.0",
    "google-auth>=2.0",
    "google-auth-oauthlib>=1.0",
]
# axe-core ベース a11y スキャン: scripts/run_a11y_scan.py が使用
a11y = [
    "axe-playwright-python>=0.1.4",
]
# 開発用: pure function の単体テスト (uv run --extra dev pytest)
dev = [
    "pytest>=7.0",
]

[project.scripts]
scenario-test = "scenario_test.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scenario_test"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
